requirements.txt
Django==1.6.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-tastypie==0.11.0
django-toolbelt==0.0.1
gunicorn==18.0
ipdb==0.8
ipython==1.2.1
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
static==1.0.2
wsgiref==0.1.2
